requirements.txt
pymupdf             
pandas              
openpyxl  
numpy       
pytesseract 
pillow    
pdfplumber
img2table
pdfminer.six 
scikit-learn
ocrmypdf
